did changes to guesstimate function

+++ src_charm_driver/utility/util.C
@@ -2287,7 +2287,7 @@ void Config::guesstimateParms(int natm_nl, int sizez){
 	//	  sGrainSize/2;
 	int numGrains = nstates/sGrainSize;
 	numGrains*=numGrains;
-	nchareG = gExpandFact*(double)low_x_size;
+	nchareG = (int)(gExpandFact*(double)low_x_size);
 	numChunks=numPes/(nchareG*numGrains);
 	if(numChunks<1)
 	  numChunks=1;
